pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "desi-cmb-fli"
version = "0.1.0"
description = "Field-level inference pipeline for DESI × CMB lensing"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Jonathan Hawla", email = "jonathan.hawla@example.com"}]
keywords = ["cosmology", "field-level inference", "DESI", "CMB", "lensing", "astronomy"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Astronomy",
    "Topic :: Scientific/Engineering :: Physics",
]
dependencies = [
  "numpy",
  "scipy",
  "matplotlib",
  "healpy",
  "astropy",
  "fitsio",
  "pyyaml",
  "rich",
  "tqdm",
  "xarray",
  "hydra-core",
  "omegaconf",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-cov",
  "black",
  "ruff",
  "mypy",
  "pre-commit",
]
docs = [
  "mkdocs",
  "mkdocs-material",
  "pymdown-extensions",
]
jax = [
  "jax",
  "jaxlib",
  "flax",
  "optax",
  "numpyro",
  "chex",
  "jax-cosmo",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "C4"]
ignore = ["E203", "E266", "E501"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q --color=yes"
